fix: Phase 3 batch 1-4 — dangerous conversion warnings in small/medium modules
Fix -Wshorten-64-to-32, -Wfloat-conversion, -Wimplicit-float-conversion
across wwutil, wwsaveload, wwtranslatedb, BinkMovie, SControl, wwnet,
WWAudio, compat, wwphys, WWMath (46 files, ~180 unique sites).

Notable fixes:
- netutil.cpp: u_long* → uint32_t* for IPv4 address read (was overreading
  8 bytes from a 4-byte field on arm64 — real 64-bit bug)
- timeGetTime() call sites: narrowed to uint32_t (timer wraps at 32 bits)
- MILES_HANDLE/HSAMPLE/H3DSAMPLE handle storage: unsigned long → uint32_t
  (handles are 1-based indices ≤ 64, always fit in 32 bits)
- audiblesound.h: float→unsigned long for playback position (intentional
  normalized-float × duration-ms arithmetic)
- chunkio.h: WRITE_MICRO_CHUNK_STRING macro size cast (fixes all call sites)
- quat.cpp: sqrt→sqrtf, 0.5→0.5f etc. (eliminates double promotion)
- WWMath tri.h: float→bool converted to \!= 0.0f comparisons
- All float→int conversions use static_cast<int32_t> (explicit truncation)
- All ssize_t/ptrdiff_t→int use static_cast<int32_t> (bounded socket bufs)

Warnings: 24,693 → 19,521 (-5,172)

+++ original/Code/BinkMovie/BINKMovie.cpp
@@ -279,14 +279,14 @@ void BINKMovieClass::Render()
 				D3DSURFACE_DESC d3d_surf_desc;
 				D3DLOCKED_RECT locked_rect;
 
-				DX8_ErrorCode(d3d_texture->GetLevelDesc(0, &d3d_surf_desc));
+				DX8_ErrorCode(static_cast<uint32_t>(d3d_texture->GetLevelDesc(0, &d3d_surf_desc)));
 
 				RECT rect;
 				rect.left = 0;
 				rect.top = 0;
 				rect.right = w;
 				rect.bottom = h;
-				DX8_ErrorCode(d3d_texture->LockRect(0,&locked_rect,&rect,0));
+				DX8_ErrorCode(static_cast<uint32_t>(d3d_texture->LockRect(0,&locked_rect,&rect,0)));
 
 				for (unsigned y = 0; y < h; ++y) {
 					unsigned char* dest = (unsigned char*)locked_rect.pBits + y * locked_rect.Pitch;
@@ -294,7 +294,7 @@ void BINKMovieClass::Render()
 					cur_tex_ptr += Bink->Width * 2;
 				}
 
-				DX8_ErrorCode(d3d_texture->UnlockRect(0));
+				DX8_ErrorCode(static_cast<uint32_t>(d3d_texture->UnlockRect(0)));
 			}
 		}
