+++ original/Code/WWAudio/AudibleSound.h
@@ -197,7 +197,7 @@ class AudibleSoundClass : public SoundSceneObjClass
 		//
 		virtual unsigned long	Get_Duration (void) const								{ return m_Length; }
 		virtual unsigned long	Get_Play_Position (void) const						{ return m_CurrentPosition; }
-		virtual void				Set_Play_Position (float position)					{ Seek (position * m_Length); }
+		virtual void				Set_Play_Position (float position)					{ Seek (static_cast<unsigned long>(position * m_Length)); }
 		virtual void				Set_Play_Position (unsigned long milliseconds)	{ Seek (milliseconds); }
 
 		virtual void				Set_Start_Offset (float offset)						{ m_StartOffset = offset; }
