+++ original/Code/wwlib/chunkio.h
@@ -281,7 +281,7 @@ inline void write_micro_chunk_value(ChunkSaveClass& csave, const T& var) {
 
 #define WRITE_MICRO_CHUNK_STRING(csave,id,var) { \
 	csave.Begin_Micro_Chunk(id); \
-	csave.Write(var, strlen(var) + 1); \
+	csave.Write(var, static_cast<uint32_t>(strlen(var) + 1)); \
 	csave.End_Micro_Chunk(); }
 
 #define WRITE_MICRO_CHUNK_WWSTRING(csave,id,var) { \
